backend/requirements.txt
livekit-agents[silero,turn-detector]>=1.4.0
livekit-plugins-noise-cancellation>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
